pyproject.toml
[project]
name = "homelab-pyinfra"
version = "0.1.0"
description = "Infrastructure as Code for homelab management using PyInfra"
readme = "README.md"
requires-python = ">=3.14"
authors = [
    {name = "homelab-admin", email = "admin@mental404.com"}
]
license = {text = "MIT"}
keywords = ["homelab", "infrastructure", "automation", "pyinfra"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.14",
]

dependencies = [
    "pyinfra>=3.6.1",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-testinfra>=10.2.2",
    "pytest-cov>=4.1.0",
    "pytest-html>=4.1.1",
    "pytest-rerunfailures>=13.0",
    "ruff>=0.2.0",
    "black>=24.0.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["deploy", "deploy.py"]

[tool.ruff]
line-length = 100
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.black]
line-length = 100
target-version = ["py314"]

[tool.mypy]
python_version = "3.14"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = ["-v", "--strict-markers", "--tb=short"]
markers = [
    "slow: marks tests as slow",
    "docker: marks tests requiring Docker",
    "network: marks tests requiring network access",
]
